FIX counterbalancing of stims
In the original stimuli masterlist, there was an error in how the
nonword was distributed over the priming trials. It varied between no
nonword being present and the nonword appearing on the third word. This
technically should not have affected the distribution on nonwords on
each word region but across the two lists, there was a difference where
you consistently saw a nonword after the priming condition versus a
word.