Update pca with new specs
window from -100ms to 30 ms. using Evoked instead of Epochs based on
conversation with Alex on June 8th. Possibility that compute of
projector on the epochs is finding inter-trial variation, which we are
not interested in. Reducing time window also improves sensitivity to
the saccade effect of interest by remove brain components that were
associated with the visual response (around 100ms). Old time window was
-100 to 100 ms.