ENH epochs, plots
Create a common epochs file with all the events for the experiment.
This is currently has redundant events because some triggers here point
the same event. This is being worked on right now with mne-python, and
may be resolved. If not, the file will be slightly bloated.

Enhance the plot to have a single figure containing the waveform of the
PCA’d data, and the PC topography.
Use a common epochs file.