ENH add make_raw, rERF
add a script for creating a filtered raw. don’t necessarily like it
because i was working primarily on the epochs files and this now adds a
redundant copy but the raw is necessary for the rER[P/F]s.
It seems that if the events are overlapping this needs to be done to
correct for that. I need to figure out why this is the case, i.e., I
need to read the Smith paper part 2.

Add the rERF to the priming_sensor_analysis.

To-Do: add it to the fixation_times. Figure out what stats can be done
on these evokeds.