Update priming_sensor_analysis
This now includes rERF implementation and a spatio-temporal cluster
test. still wip.
The GAT class is currently implemented as well. What’s needed is an
overall score taking all time and sensors into account to make a
prediction.
The pca now has an explained variance included in plots.

The basis of analyses now rely on working from the raw instead of the
epochs. It was seen as the most practical option.

What’s needed now is an analysis on the behavioral data of the two
conditions at the bare minimum. Then we can start looking at the brain
data again.