Update analyses
For word-nonword, and priming, the epochs object must be reduced to the
comparisons of interests. After a few additional hours of computation
time, and horribly rendered result files, I found that all the
available events were labeled and being predicted. NO BUENO.

To be improved, we need to find a good way to save objects in mne.

On a more positive note, separating the analysis from the plotting
makes the code easier to read, and it makes redoing plots easier.

Minor revisions to config, there was a redundant event_id dict, and the
fixation id causes problems with the evts when it is not a specified
value.