Coregistration pipeline
There are currently two ways to get behavioral responses, one from the
region that is dumped in the dat file, and the second is from the
actual fixation data, which comes from the edf.

For the MEG-eyetracking coregistration, we will be using the
first-fixation duration as a correlation to the brain activity. This
requires the edf data (pipeline b).
The design matrix now encompasses all the primes and target fixations
for a more robust predictive model. It used to be restricted to the
target region for reasons I don’t remember.