Massive code cleanup
This is the reimplementation of the pipeline from start to finish.
All the raw files were visually inspected by me, and appropriate
channel interpolation or chunk of time deleted.
The data have been filtered three four ways for comparison.
(0.03, 200), (0.03, 40), (0.51, 40), (1, 40)
Rework flow of the analysis scripts for priming and nonword/word.