completed make_trans
this script makes a copy of the template brain, fsaverage, and semi-automatically fits it to the fiducials. the transformation matrix allows for us to go from head space using the digitizer fiducials to mri space using the template fudicials. ultimate we will have a combination of transforms used for source estimation:
- the meg device to head space transform done with hpi coils in meg device space to equivalent digitizer points in head space
- head to device space computed with the trans computation in this script

Minor update to other scripts and wip requirements file