Cleanup and overhaul
At first, I was considering. using this pattern but this would require where in the file hierarchy you are and would need to modify it if the file is moved.

`root = Path(__file__).resolve().parents[5]`

Instead I'm using the Path parents and matching on the project root name `OcularLDT-project`

Simplified the code and made use of the BIDSPath